+++ src/api2/managers/wallets-manager.js
@@ -4,6 +4,8 @@ import _get from 'lodash/get'
 import { Wallet } from '../../wallet'
 import { Keypair } from '../../base/keypair'
 
+import { Signer } from './signer'
+
 import { SignersManager } from './signers-manager'
 import { ApiCaller } from '../api-caller'
 
@@ -46,6 +48,10 @@ export class WalletsManager {
       email,
       is_recovery: isRecovery
     })
+  }
+
+  getSignerRoleId () {
+    return this._apiCaller.get('/v3/key_values/signer_role:default')
   }
 
   /**
@@ -95,13 +101,18 @@ export class WalletsManager {
    *
    * @param {string} email User's email.
    * @param {string} password User's password.
-   * @param {Keypair} recoveryKeypair the keypair to later recover the account
-   * @param {string} [referrerId] public key of the referrer
+   * @param {Array} [signers] array of {@link Signer}
    *
-   * @return {Promise.<object>} User's wallet and a recovery seed.
+   * @return {Promise.<object>} User's wallet.
    */
-  async create (email, password, recoveryKeypair, referrerId = '') {
+  async createWithSigners (email, password, signers = []) {
+    signers.forEach(item => {
+      if (!(item instanceof Signer)) {
+        throw new TypeError('A signer instance expected.')
+      }
+    })
     const { data: kdfParams } = await this.getKdfParams('')
+    const { data: roleId } = await this.getSignerRoleId()
 
     const mainWallet = Wallet.generate(email)
     const encryptedMainWallet = mainWallet.encrypt(kdfParams, password)
@@ -111,14 +122,20 @@ export class WalletsManager {
       kdfParams, password
     )
 
-    const walletRecoveryKeypair = recoveryKeypair || Keypair.random()
-    const encryptedRecoveryWallet = mainWallet.encryptRecoveryData(
-      kdfParams, walletRecoveryKeypair
-    )
+    const defaultSigner = new Signer({
+      id: mainWallet.accountId,
+      roleId: roleId.value.u32,
+      weight: 1000,
+      identity: 1
+    })
+    signers.push(defaultSigner)
 
-    const accountRefferer = referrerId
-      ? { referrer: { data: { id: referrerId } } }
-      : {}
+    const relationshipsSigners = signers.map(item => {
+      return {
+        type: item.type,
+        id: item.id
+      }
+    })
 
     const response = await this._apiCaller.post('/wallets', {
       data: {
@@ -137,19 +154,15 @@ export class WalletsManager {
               id: kdfParams.id
             }
           },
-          recovery: {
-            data: {
-              type: 'recovery',
-              id: encryptedRecoveryWallet.id
-            }
-          },
           factor: {
             data: {
               type: 'password',
               id: encryptedMainWallet.id
             }
           },
-          ...accountRefferer
+          signers: {
+            data: relationshipsSigners
+          }
         }
       },
       included: [
@@ -162,15 +175,7 @@ export class WalletsManager {
             salt: encryptedSecondFactorWallet.salt
           }
         },
-        {
-          type: 'recovery',
-          id: encryptedRecoveryWallet.id,
-          attributes: {
-            account_id: encryptedRecoveryWallet.accountId,
-            keychain_data: encryptedRecoveryWallet.keychainData,
-            salt: encryptedRecoveryWallet.salt
-          }
-        }
+        ...signers
       ]
     })
 
@@ -185,9 +190,31 @@ export class WalletsManager {
 
     return {
       wallet: walletWithSession,
-      response: response,
-      recoverySeed: walletRecoveryKeypair.secret()
+      response: response
     }
+  }
+
+  /**
+   * Create a wallet.
+   *
+   * @param {string} email User's email.
+   * @param {string} password User's password.
+   * @param {Keypair} recoveryKeypair the keypair to later recover the account
+   * @param {string} [referrerId] public key of the referrer
+   *
+   * @return {Promise.<object>} User's wallet and a recovery seed.
+   */
+  async create (email, password, recoveryKeypair, referrerId = '') {
+    const walletRecoveryKeypair = recoveryKeypair || Keypair.random()
+    const recoverySigner = new Signer({
+      id: walletRecoveryKeypair.accountId(),
+      roleId: 1,
+      weight: 1000,
+      identity: 1
+    })
+    const wallet = await this.createWithSigners(email, password, [recoverySigner])
+    wallet.recoverySeed = walletRecoveryKeypair.secret()
+    return wallet
   }
 
   /**
@@ -235,6 +262,7 @@ export class WalletsManager {
    */
   async recovery (email, recoverySeed, newPassword) {
     const { data: kdfParams } = await this.getKdfParams(email, true)
+    const accountId = await this._getAccountIdByEmail(email)
 
     const recoveryWallet = Wallet.fromRecoverySeed(
       kdfParams,
@@ -243,15 +271,14 @@ export class WalletsManager {
       recoverySeed
     )
 
-    const newMainWallet = Wallet.generate(email)
+    const newMainWallet = Wallet.generate(email, accountId)
     const encryptedNewMainWallet = newMainWallet.encrypt(kdfParams, newPassword)
 
-    const newSecondFactorWallet = Wallet.generate(email)
+    const newSecondFactorWallet = Wallet.generate(email, accountId)
     const encryptedSecondFactorWallet = newSecondFactorWallet.encrypt(
       kdfParams, newPassword
     )
 
-    const accountId = await this._getAccountIdByRecoveryId(recoveryWallet.id)
     const tx = await this._signersManager.createChangeSignerTransaction({
       newPublicKey: newMainWallet.accountId,
       signingKeypair: recoveryWallet.keypair,
@@ -473,5 +500,21 @@ export class WalletsManager {
     const { data: wallet } = await this._apiCaller.get(endpoint)
 
     return wallet.accountId
+  }
+
+  /**
+   * Get user account ID by email.
+   *
+   * @param {string} email account email.
+   *
+   * @return {Promise.<string>} User's account ID.
+   */
+  async _getAccountIdByEmail (email) {
+    const { data } = await this._apiCaller.get('/identities', {
+      filter: { identifier: email },
+      page: { limit: 1 }
+    })
+
+    return _get(data[0], 'address')
   }
 }